+++ src/batch-transfers/batch-transfers.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common'
 import { BatchTransfersService } from './batch-transfers.service'
 import { BatchTransfersController } from './batch-transfers.controller'
+import { BatchTransfersSchedule } from './batch-transfers.schedule'
 import { UsersModule } from '../users/users.module'
 
 @Module({
   imports: [UsersModule],
-  providers: [BatchTransfersService],
+  providers: [BatchTransfersService, BatchTransfersSchedule],
   controllers: [BatchTransfersController],
   exports: [BatchTransfersService],
 })